Add a more complex log parsing question.
This question uses a big apache log dataset, and a more complicated
query (filter and aggregate) that is very typical of apache log
analysis.  It is a big enough problem that a user could realistically
chose to solve it in a simple python/perl/ruby script or could use a
more advanced shell pipeline (e.g. containing awk).

This is a great question for folks who will be maintaining applications.

For the time, I've chosen not to include the apache log file, as it is
rather large and I don't want to clog up the repo. Maybe git-lfs?